# server/docs/swagger.yaml
basePath: /
definitions:
...
  database.ProviderProfile:
    properties:
      createdAt:
        type: string
      manifest_url:
        description: url to manifest.json for stremio providers
        type: string
      name:
        description: profile name, eg. Performance, Quality, etc.
        type: string
      provider_id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  v1.CreateProviderRequest:
    properties:
      manifest_url:
        type: string
      name:
        type: string
    type: object
...
  v1.V1ErrorResponse:
    properties:
      error:
        example: error message
        type: string
      status:
        enum:
        - success
        - error
        example: error
        type: string
    type: object
  v1.V1SuccessResponse:
    properties:
      data: {}
      status:
        enum:
        - success
        - error
        example: success
        type: string
    type: object
...
host: localhost:2323
info:
  contact: {}
  description: |-
    V1 Hound Media Server API Documentation
    While in beta, please expect breaking api changes in short/no notice
  title: Hound API Documentation V1
  version: beta-0.1.0
paths:
...
  /api/v1/provider_profiles:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/database.ProviderProfile'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get all provider profiles
      tags:
      - Provider Profiles
    post:
      consumes:
      - application/json
      parameters:
      - description: Provider Profile
        in: body
        name: provider_profile
        required: true
        schema:
          $ref: '#/definitions/v1.CreateProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/database.ProviderProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Create a provider profile
      tags:
      - Provider Profiles
  /api/v1/provider_profiles/{providerID}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Provider ID
        in: path
        name: providerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Delete a provider profile
      tags:
      - Provider Profiles
...
  /api/v1/stream/{encodedString}:
    get:
      consumes:
      - application/json
      description: A streamable link for a video defined by the encodedString
      parameters:
      - description: Encoded Stream Details
        in: path
        name: encodedString
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Stream Video
      tags:
      - Stream
  /api/v1/torrent/{encodedString}:
    post:
      consumes:
      - application/json
      description: |-
        Adds a p2p torrent to the server for streaming/download.
        Not strictly necessary, as calling stream directly invokes this. May be deprecated in the future.
      parameters:
      - description: Encoded Stream Details
        in: path
        name: encodedString
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Add Torrent
      tags:
      - Stream
...
swagger: "2.0"